rtoot: Collect Mastodon data from R

rtoots provides a range of functions to interact with the Mastodon API. It has five main groups of functions:

get_account_<type>() to get account level data

get_timeline_<type>() to get series of statuses

get_instance_<type>() get information for Mastodon instances

stream_<type>() to stream statuses from different timeline

post_<type>() to push content to Mastodon from R

(Some auxiliary functions do not follow this naming scheme)

Get started by reading vignette("rtoot"). and browse all vignettes with browseVignettes(package = "rtoot")

See also

Author

Maintainer: David Schoch david@schochastics.net (ORCID)

Authors:

Chung-hong Chan chainsawtiney@gmail.com (ORCID)

Other contributors:

Johannes Gruber JohannesB.Gruber@gmail.com (ORCID) [contributor]

Tim Schatto-Eckrodt kudusch@posteo.de (ORCID) [contributor]